# Remove scalar sha3 amd64 assembly