… internal/sha3/keccakf_amd64.go
… internal/sha3/keccakf_amd64.s